routing added on main.dart
Following the criticism the routing is added. As much as I wanted to do it dynamically taking the pose-id that seem to be impossible for routing{} and can possibly achieved via external libraries to handle it.

Instead when a pose is clicked the url adds /pose
through setUrlStrategy(PathUrlStrategy()); in void main() with import of import 'package:flutter_web_plugins/flutter_web_plugins.dart';
the # that divides the url eg  http://localhost:52744/#/pose
so the result is http://localhost:52744/pose

That is the best I can do with the current situation and all of the other options seem to be doing the app stateless and hardcoding the data then passing the id in PosePage as it was at the beginning.